==> Binary_classifier/2.Cleaning.ipynb <==
# Cleaning the data


# 1)- Importing key modules

# support both Python 2 and Python 3 with minimal overhead.
from __future__ import absolute_import, division, print_function

# I am an engineer. I care only about error not warning. So, let's be maverick and ignore warnings.
import warnings
warnings.filterwarnings('ignore')

import re    # for regular expressions 
import nltk  # for text manipulation 
import string 
import numpy as np 
import pandas as pd 

#For Visuals
import matplotlib.pyplot as plt
import seaborn as sns
sns.set()
from matplotlib import rcParams
rcParams['figure.figsize'] = 11, 8
%config InlineBackend.figure_format = 'svg'
%matplotlib inline

%reload_ext version_information
%version_information pandas,numpy, nltk, seaborn, matplotlib

# 2)- Loading Dataset

data=pd.read_csv('final_data_w_classes_only_contracts.csv',index_col=[0])

data.shape

data.head()

data.tail()

data['class'].value_counts()

**0 means other text and 1 means Deliverable and Obligations**

# And Deliverables^Objective class look like
data[data["class"]==1.0].head(10)

# let's see how Other type classlooks like
data[data["class"]==0.0].head(10)

len(str(data['text']))

### check missing values

# Checking rows now
def summary_missing(dataset):
    n_miss = dataset.isnull().sum()
    n_obs = dataset.shape[0]
    n_miss_per = n_miss/n_obs*100
    n_miss_tbl = pd.concat([n_miss, n_miss_per], axis = 1).sort_values(1, ascending = False).round(1)
    n_miss_tbl = n_miss_tbl[n_miss_tbl[1] != 0]
    print('No. of fields: ', dataset.shape[0])
    print('No. of missing fields: ', n_miss_tbl.shape[0])
    n_miss_tbl = n_miss_tbl.rename(columns = {0:'No. of mising Value', 1:'%age of missing Value'})
    return n_miss_tbl

summary_missing(data)

data_class=data['class']

What to do with them? Delete them or ffill?

data_class.unique()

# where does missing value exist!!!

data_class.loc[data_class.isnull()]

# filling values

data = data.fillna(method="ffill")

summary_missing(data)

# 3)- Text Cleaning

### 3.1)-remove unwanted text patterns

def remove_pattern(input_txt, pattern):
    r = re.findall(pattern, input_txt)
    for i in r:
        input_txt = re.sub(i, '', input_txt)
    return input_txt

What are some unwanted patterns. Let's say we have @ now

data['clean'] = np.vectorize(remove_pattern)(data['text'], "@[\w]*") 
data.head()

data.text[0]

data.clean[0]

### 3.2)-Removing Punctuations, Numbers, and Special Characters

data['clean'] = data['clean'].str.replace("[^a-zA-Z#]", " ") 
data.head()

data.text[0]

data.clean[0]

### 3.3)-Text Normalization

tokenized_text = data['clean'].apply(lambda x: x.split()) # tokenizing 
tokenized_text.head()

print(tokenized_text[0])

print(tokenized_text[1])

nltk.download('wordnet')

from nltk.stem import WordNetLemmatizer
lemm = WordNetLemmatizer()

tokenized_text = tokenized_text.apply(lambda x: [lemm.lemmatize(i) for i in x])

print(tokenized_text[0])

print(tokenized_text[5])

# stitch these tokens back together.

for i in range(len(tokenized_text)):
    tokenized_text[i] = ' '.join(tokenized_text[i])    
data['clean'] = tokenized_text

data.text[3]

data.clean[3]

data.head()

### All in one quick command

from nltk.corpus import stopwords
clean_text = [] # defining corpus
for i in range(0, len(data['text'])): # giving range of values from 0 to 1000
    processed_text = re.sub('[^a-zA-Z]', ' ', data['text'][i]) # using i for all values instead of 1 column
    processed_text = processed_text.lower()
    processed_text = processed_text.split()
    lemm = WordNetLemmatizer()
    stopword_set = set(stopwords.words('english'))
    processed_text = [lemm.lemmatize(word) for word in processed_text if not word in stopword_set]
    processed_text = ' '.join(processed_text )
    clean_text.append(processed_text) # finally attach all these cleaned values to corpus directory

# Remove numbers
clean_text = [word for word in clean_text if not word.isnumeric()]

type(clean_text)

# convert to series to check words
text_corpus = pd.Series(clean_text)

data['clean2'] = text_corpus
data.head()

data.clean2[2]

data.clean2[0]

data.clean2[3]

data.text[3]

data.clean[3]

We can see that "clean" column provides better cleaning practices compared to other two. For example, E.ON is detected as "e" in clean2 and "E ON" in clean. Reason is that I did use lower case and hence, "on" has been taken as stopword and got removed.

# save data
import pickle
data.to_pickle('file_clean.pkl')

data_pickle=pd.read_pickle('file_clean.pkl')
data_pickle.head()
