==> Binary_classifier/4.Modeling.ipynb <==
# Modeling

Before modeling, there will be EDA, Cleaning and vectorization procedures.

# 1)- Importing key modules

#support both Python 2 and Python 3 with minimal overhead.
from __future__ import absolute_import, division, print_function
import warnings
warnings.filterwarnings('ignore')

import re    # for regular expressions 
import nltk  # for text manipulation 
import string 
import numpy as np 
import pickle
import pandas as pd 

#For Visuals
import matplotlib.pyplot as plt
import seaborn as sns
sns.set()
from matplotlib import rcParams
rcParams['figure.figsize'] = 11, 8
%config InlineBackend.figure_format = 'svg'
%matplotlib inline

# 2)- Loading Dataset

data=pd.read_pickle('file_clean.pkl')

data.shape

data.head()

data.tail()

data['class'].value_counts()

**0 means other text and 1 means Deliverable and Obligations**

# 3)- Model

- Logistic Regression
- Support Vector Machine
- Naives Bayes
- RandomForest
- XGBoost


# loading pickle files before we start training

bow=pd.read_pickle('bow_model.pkl')
tfidf=pd.read_pickle('tfidf_model.pkl')
wordvec_df=pd.read_pickle('word2vec_model.pkl')
docvec_df=pd.read_pickle('doc2vec_model.pkl')

### 3.1)- Logistic Regression Model

from sklearn.linear_model import LogisticRegression 
from sklearn.model_selection import train_test_split 
from sklearn.metrics import f1_score, accuracy_score, roc_curve,roc_auc_score,confusion_matrix, classification_report

#### 3.1.1 Logistic Regression using Bag-of-Words Features

X=bow
y=data['class']

print(X.shape)
print(y.shape)

# splitting data into training and validation set
xtrain_bow, xvalid_bow, ytrain, yvalid = train_test_split(bow, y,random_state=42,test_size=0.2)

print(xtrain_bow.shape)
print(xvalid_bow.shape)
print(ytrain.shape)
print(yvalid.shape)

lreg = LogisticRegression(solver='liblinear')

# training the model
lreg.fit(xtrain_bow, ytrain)

# predicting on the validation set
prediction = lreg.predict_proba(xvalid_bow)
prediction

# prediction over classes

prediction_class=lreg.predict(xvalid_bow)
prediction_class

accuracy_score(yvalid, prediction_class)

roc_auc_score(yvalid, prediction_class)

confusion_matrix(yvalid, prediction_class)

print(classification_report(yvalid, prediction_class))

**Changing threshold**

# if prediction is greater than or equal to 0.3 than 1 else 0
prediction_int = prediction[:,1] >= 0.3
prediction_int

prediction_int = prediction_int.astype(np.int)
prediction_int

# calculating f1 score for the validation set
f1_score(yvalid, prediction_int)

# calculating accuracy score for the validation set
accuracy_score(yvalid, prediction_int)

# roc-auc score
roc_auc_score(yvalid, prediction_int)

confusion_matrix(yvalid, prediction_int)

print(classification_report(yvalid, prediction_int))

By changing threshold, we have seen a decrease in accuracy but, we have got more precise results for class 1 i.e O&B. So, it is trade-off between accuracy and precision.

For this excercise, we are more interested in Class 1 precised results so, 0.3 threshold is used for rest of notebook.

#### 3.1.2 Logistic Regression using TF-IDF Features

X=tfidf
y=data['class']

print(X.shape)
print(y.shape)

# splitting data into training and validation set
xtrain_tfidf, xvalid_tfidf, ytrain, yvalid = train_test_split(tfidf, y,random_state=42,test_size=0.2)

print(xtrain_tfidf.shape)
print(xvalid_tfidf.shape)
print(ytrain.shape)
print(yvalid.shape)

lreg = LogisticRegression(solver='liblinear')

# training the model
lreg.fit(xtrain_tfidf, ytrain)

# predicting on the validation set
prediction = lreg.predict_proba(xvalid_tfidf)
prediction

# Same rule : if prediction is greater than or equal to 0.3 than 1 else 0
prediction_int = prediction[:,1] >= 0.3
prediction_int

prediction_int = prediction_int.astype(np.int)
prediction_int

accuracy_score(yvalid, prediction_int)

# calculating f1 score for the validation set
f1_score(yvalid, prediction_int)

confusion_matrix(yvalid, prediction_int)

print(classification_report(yvalid, prediction_int))

#### 3.1.3 Logistic Regression using Word2Vec Features

X=wordvec_df
y=data['class']

print(X.shape)
print(y.shape)

# splitting data into training and validation set
xtrain_word2vec, xvalid_word2vec, ytrain, yvalid = train_test_split(X, y,random_state=42,test_size=0.2)

print(xtrain_word2vec.shape)
print(xvalid_word2vec.shape)
print(ytrain.shape)
print(yvalid.shape)

# training the model
lreg.fit(xtrain_word2vec, ytrain)

# predicting on the validation set
prediction = lreg.predict_proba(xvalid_word2vec)
prediction

prediction_int = prediction[:,1] >= 0.3
prediction_int

prediction_int = prediction_int.astype(np.int)
prediction_int

f1_score(yvalid, prediction_int)

accuracy_score(yvalid, prediction_int)

confusion_matrix(yvalid, prediction_int)

print(classification_report(yvalid, prediction_int))

#### 3.1.4 Logistic Regression using Doc2Vec Features

X=docvec_df
y=data['class']

print(X.shape)
print(y.shape)

# splitting data into training and validation set
xtrain_doc2vec, xvalid_doc2vec, ytrain, yvalid = train_test_split(X, y,random_state=42,test_size=0.2)

print(xtrain_doc2vec.shape)
print(xvalid_doc2vec.shape)
print(ytrain.shape)
print(yvalid.shape)

# training the model
lreg.fit(xtrain_doc2vec, ytrain)

# predicting on the validation set
prediction = lreg.predict_proba(xvalid_doc2vec)
prediction

prediction_int = prediction[:,1] >= 0.3
prediction_int

prediction_int = prediction_int.astype(np.int)
prediction_int

f1_score(yvalid, prediction_int)

accuracy_score(yvalid, prediction_int)

confusion_matrix(yvalid, prediction_int)

print(classification_report(yvalid, prediction_int))

For a simpler dataset, bag of words is giving best score. Howver, we can see that tfidf and word2vec performs very well i.e 0.94 score for class 1 and very bad for class 0. So, we may use them if we are interested in more precise results for class 1.

### 3.2)-Support Vector Machine (SVM)

from sklearn import svm

#### 3.2.1 SVM using Bag-of-Words Features

svc = svm.SVC(kernel='linear', C=1, probability=True).fit(xtrain_bow, ytrain)

prediction = svc.predict_proba(xvalid_bow)
prediction

prediction_int = prediction[:,1] >= 0.3
prediction_int

prediction_int = prediction_int.astype(np.int)
prediction_int

accuracy_score(yvalid, prediction_int)

f1_score(yvalid, prediction_int)

confusion_matrix(yvalid, prediction_int)

print(classification_report(yvalid, prediction_int))

#### 3.2.2 SVM using TF-IDF Features

svc = svm.SVC(kernel='linear',C=1, probability=True).fit(xtrain_tfidf, ytrain)

prediction = svc.predict_proba(xvalid_tfidf)
prediction

prediction_int = prediction[:,1] >= 0.3
prediction_int

prediction_int = prediction_int.astype(np.int)
prediction_int

accuracy_score(yvalid, prediction_int)

f1_score(yvalid, prediction_int)

confusion_matrix(yvalid, prediction_int)

print(classification_report(yvalid, prediction_int))

#### 3.2.3 SVM using word2vec Features

svc = svm.SVC(kernel='linear', C=1, probability=True).fit(xtrain_word2vec, ytrain)

prediction = svc.predict_proba(xvalid_word2vec)
prediction

prediction_int = prediction[:,1] >= 0.3
prediction_int

prediction_int = prediction_int.astype(np.int)
prediction_int

accuracy_score(yvalid, prediction_int)

f1_score(yvalid, prediction_int)

print(confusion_matrix(yvalid, prediction_int))

print(classification_report(yvalid, prediction_int))

#### 3.2.4 SVM using doc2vec Features

svc = svm.SVC(kernel='linear', C=1, probability=True).fit(xtrain_doc2vec, ytrain)

prediction = svc.predict_proba(xvalid_doc2vec)
prediction

prediction_int = prediction[:,1] >= 0.3
prediction_int

prediction_int = prediction_int.astype(np.int)
prediction_int

accuracy_score(yvalid, prediction_int)

f1_score(yvalid, prediction_int)

print(confusion_matrix(yvalid, prediction_int))

print(classification_report(yvalid, prediction_int))

### 3.3)- Naive Bayes

from sklearn.naive_bayes import MultinomialNB

#### 3.3.1 Naive Bayes using Bag-of-Words 

mnb = MultinomialNB()
mnb.fit(xtrain_bow, ytrain)

# naive bayes in theory does not work well with predict_probab so, we shall use prediction directly
prediction = mnb.predict_proba(xvalid_bow)
prediction

prediction_int = prediction[:,1] >= 0.3
prediction_int

prediction_int = prediction_int.astype(np.int)
prediction_int

accuracy_score(yvalid, prediction_int)

f1_score(yvalid, prediction_int)

print(confusion_matrix(yvalid, prediction_int))

print(classification_report(yvalid, prediction_int))

Not too good

#### 3.3.2 Naive Bayes using TF-IDF

mnb = MultinomialNB()
mnb.fit(xtrain_tfidf, ytrain)

prediction = mnb.predict_proba(xvalid_tfidf)
prediction

prediction_int = prediction[:,1] >= 0.3
prediction_int

prediction_int = prediction_int.astype(np.int)
prediction_int

accuracy_score(yvalid, prediction_int)

f1_score(yvalid, prediction_int)

print(confusion_matrix(yvalid, prediction_int))

print(classification_report(yvalid, prediction_int))

#### 3.3.3 Naive Bayes using word2vec

As for Naive Bayes, Input X must be non-negative. In case of Word2vec and Doc2vec, Naive Bayes is not ideal. So we will not use it. It is not giving any exceptional results anyway.

For results, we have already have better models' outcome.

#### 3.3.4 Naive Bayes using doc2vec

### 3.4 Random Forest

from sklearn.ensemble import RandomForestClassifier

#### 3.4.1-RF with Bag-of-Words Features

rf = RandomForestClassifier(n_estimators=400, random_state=11).fit(xtrain_bow, ytrain)

prediction = rf.predict_proba(xvalid_bow)
prediction

prediction_int = prediction[:,1] >= 0.3
prediction_int

prediction_int = prediction_int.astype(np.int)
prediction_int

f1_score(yvalid, prediction_int)

accuracy_score(yvalid, prediction_int)

print(confusion_matrix(yvalid, prediction_int))

print(classification_report(yvalid, prediction_int))

#### 3.4.2-RF with TF-IDF Features

rf = RandomForestClassifier(n_estimators=400, random_state=11).fit(xtrain_tfidf, ytrain)

prediction = rf.predict_proba(xvalid_tfidf)
prediction

prediction_int = prediction[:,1] >= 0.3
prediction_int

prediction_int = prediction_int.astype(np.int)
prediction_int

accuracy_score(yvalid, prediction_int)

f1_score(yvalid, prediction_int)

print(confusion_matrix(yvalid, prediction_int))

print(classification_report(yvalid, prediction_int))

#### 3.4.3-RF with word2vec Features

print(xtrain_word2vec.shape)
print(xvalid_word2vec.shape)

rf = RandomForestClassifier(n_estimators=400, random_state=11).fit(xtrain_word2vec, ytrain)

prediction= rf.predict_proba(xvalid_word2vec)
prediction

prediction_int = prediction[:,1] >= 0.3
prediction_int

prediction_int = prediction_int.astype(np.int)
prediction_int

accuracy_score(yvalid, prediction_int)

f1_score(yvalid, prediction_int)

print(confusion_matrix(yvalid, prediction_int))

print(classification_report(yvalid, prediction_int))

#### 3.4.4-RF with doc2vec Features

print(xtrain_doc2vec.shape)
print(xvalid_doc2vec.shape)

rf = RandomForestClassifier(n_estimators=400, random_state=11).fit(xtrain_doc2vec, ytrain)

prediction= rf.predict_proba(xvalid_doc2vec)
prediction

prediction_int = prediction[:,1] >= 0.3
prediction_int

prediction_int = prediction_int.astype(np.int)
prediction_int

accuracy_score(yvalid, prediction_int)

f1_score(yvalid, prediction_int)

print(confusion_matrix(yvalid, prediction_int))

print(classification_report(yvalid, prediction_int))

### 3.5)-XGBoost 

Extreme Gradient Boosting (xgboost) is an advanced implementation of gradient boosting algorithm. It has both linear model solver and tree learning algorithms. Its ability to do parallel computation on a single machine makes it extremely fast. It also has additional features for doing cross validation and finding important variables. There are many parameters which need to be controlled to optimize the model.

Some key benefits of XGBoost are:

- Regularization - helps in reducing overfitting
- Parallel Processing - XGBoost implements parallel processing and is blazingly faster as compared to GBM.
- Handling Missing Values - It has an in-built routine to handle missing values.
- Built-in Cross-Validation - allows user to run a cross-validation at each iteration of the boosting process

from xgboost import XGBClassifier

#### 3.5.1) XGBoost using bag of words features

xgb_model = XGBClassifier(max_depth=6, n_estimators=1000).fit(xtrain_bow, ytrain)

prediction = xgb_model.predict_proba(xvalid_bow)
prediction 

prediction_int = prediction[:,1] >= 0.3
prediction_int

prediction_int = prediction_int.astype(np.int)
prediction_int

accuracy_score(yvalid, prediction_int)

f1_score(yvalid, prediction_int)

print(confusion_matrix(yvalid, prediction_int))

print(classification_report(yvalid, prediction_int))

#### 3.5.2) XGBoost using tfidffeatures

xgb_model = XGBClassifier(max_depth=6, n_estimators=1000).fit(xtrain_tfidf, ytrain)

prediction = xgb_model.predict_proba(xvalid_tfidf)
prediction 

prediction_int = prediction[:,1] >= 0.3
prediction_int

prediction_int = prediction_int.astype(np.int)
prediction_int

accuracy_score(yvalid, prediction_int)

f1_score(yvalid, prediction_int)

print(confusion_matrix(yvalid, prediction_int))

print(classification_report(yvalid, prediction_int))

#### 3.5.3) XGBoost using word2vecfeatures

xgb_model = XGBClassifier(max_depth=6, n_estimators=1000).fit(xtrain_word2vec, ytrain)

prediction = xgb_model.predict_proba(xvalid_word2vec)
prediction 

prediction_int = prediction[:,1] >= 0.3
prediction_int

prediction_int = prediction_int.astype(np.int)
prediction_int

accuracy_score(yvalid, prediction_int)

f1_score(yvalid, prediction_int)

print(confusion_matrix(yvalid, prediction_int))

print(classification_report(yvalid, prediction_int))

#### 3.5.4) XGBoost using doc2vec features

xgb_model = XGBClassifier(max_depth=6, n_estimators=1000).fit(xtrain_doc2vec, ytrain)

prediction = xgb_model.predict_proba(xvalid_doc2vec)
prediction 

prediction_int = prediction[:,1] >= 0.3
prediction_int

prediction_int = prediction_int.astype(np.int)
prediction_int

accuracy_score(yvalid, prediction_int)

f1_score(yvalid, prediction_int)

print(confusion_matrix(yvalid, prediction_int))

print(classification_report(yvalid, prediction_int))

**XGBoost using word2vec gives us the best results with our given matrics shown in classification report in sect 6.5.3**

xgb_model_best = XGBClassifier(max_depth=6, n_estimators=1000).fit(xtrain_word2vec, ytrain)

# save model
filename = 'finalized_model.sav'
pickle.dump(xgb_model_best, open(filename, 'wb'))

# load the model from disk
loaded_model = pickle.load(open(filename, 'rb'))

prediction = loaded_model.predict_proba(xvalid_word2vec)
prediction 

# for standard threshold 0.5
prediction_class = prediction[:,1] >= 0.5
prediction_class

prediction_int = prediction_class.astype(np.int)
prediction_int

accuracy_score(yvalid, prediction_int)

f1_score(yvalid, prediction_int)

print(classification_report(yvalid, prediction_int))

# for threshold 0.5

prediction_class = prediction[:,1] >= 0.3
prediction_class

prediction_int = prediction_class.astype(np.int)
prediction_int

accuracy_score(yvalid, prediction_int)

f1_score(yvalid, prediction_int)

print(classification_report(yvalid, prediction_int))

- Why XBBoost is most optimal model ?

Reason is that it maintains a good score of accuracy and f1 for both levels of threshold i.e 0.3 and 0.5. Other models like Random Forest does not handle imbalance class very well. That's Boosting algorithms sometimes outperform ensemble learning methods and this is one of those examples.
