==> Binary_classifier/5.Model-tune.ipynb <==
# Model Tunning

We have already chosen best model. It's time to tune it and get best out of it

# 1)- Importing key modules

#support both Python 2 and Python 3 with minimal overhead.
from __future__ import absolute_import, division, print_function
import warnings
warnings.filterwarnings('ignore')

import re    # for regular expressions 
import nltk  # for text manipulation 
import string 
import numpy as np 
import pickle
import pandas as pd 

#For Visuals
import matplotlib.pyplot as plt
import seaborn as sns
sns.set()
from matplotlib import rcParams
rcParams['figure.figsize'] = 11, 8
%config InlineBackend.figure_format = 'svg'
%matplotlib inline

# For model scores and tunning

from sklearn.model_selection import train_test_split 
from sklearn.metrics import f1_score, accuracy_score, roc_curve,roc_auc_score,confusion_matrix, classification_report

# 2)- Loading Dataset & trained model

data=pd.read_pickle('file_clean.pkl')

### Load pre-processed word2vec model

from clean and vectorization process

wordvec_df=pd.read_pickle('word2vec_model.pkl')

X=wordvec_df
y=data['class']

print(X.shape)
print(y.shape)

# splitting data into training and validation set
xtrain_word2vec, xvalid_word2vec, ytrain, yvalid = train_test_split(X, y,random_state=42,test_size=0.2)

print(xtrain_word2vec.shape)
print(xvalid_word2vec.shape)
print(ytrain.shape)
print(yvalid.shape)

### Loading saved model

filename = 'finalized_model.sav'

# load the model from disk
loaded_model = pickle.load(open(filename, 'rb'))

### Evaluation of results

prediction = loaded_model.predict_proba(xvalid_word2vec)
prediction 

# for standard threshold 0.5
prediction_class = prediction[:,1] >= 0.5
prediction_class

prediction_int = prediction_class.astype(np.int)
prediction_int

accuracy_score(yvalid, prediction_int)

f1_score(yvalid, prediction_int)

print(classification_report(yvalid, prediction_int))

# for threshold 0.3

prediction_class = prediction[:,1] >= 0.3
prediction_class

prediction_int = prediction_class.astype(np.int)
prediction_int

accuracy_score(yvalid, prediction_int)

f1_score(yvalid, prediction_int)

print(classification_report(yvalid, prediction_int))

- F1 score has remained same for class 1 even though we have changed threshold.

- If we check "recall" score then there is an imporvement for class 1 from 0.83 to 0.86. So 3% of improvement is really significant.

Let's see if we can get any better.

# 3) FineTuning XGBoost + Word2Vec

import xgboost as xgb

### 3.1)- Creating DMatrices

Here we will use DMatrices. A DMatrix can contain both the features and the target.

dtrain = xgb.DMatrix(xtrain_word2vec, label=ytrain)

dvalid = xgb.DMatrix(xvalid_word2vec, label=yvalid)

# Parameters that we are going to tune 
params = {
    'objective':'binary:logistic',
    'max_depth':6,
    'min_child_weight': 1,
    'eta':.3,
    'subsample': 1,
    'colsample_bytree': 1
 }

### 3.2)- Scoring matric

So far, we have chosen vectorization method(word2vec), model(XGBoost) and it's time to choose evaluation matrix i.e f1=score as it is a binary classification problem

def custom_eval(preds, dtrain):
    labels = dtrain.get_label().astype(np.int)
    preds = (preds >= 0.3).astype(np.int)
    return [('f1_score', f1_score(labels, preds))]

# 4)- GridSearchCV

General Approach for Parameter Tuning

- Choose a relatively high learning rate. Usually a learning rate of 0.3 is used at this stage.
- Tune tree-specific parameters such as max_depth, min_child_weight, subsample, colsample_bytree keeping the learning rate fixed.
- Tune the learning rate.
- Finally tune gamma to avoid overfitting.

### 4.1)- Tuning max_depth and min_child_weight

gridsearch_params = [
    (max_depth, min_child_weight)
    for max_depth in range(6,10)
     for min_child_weight in range(5,8)
 ]
max_f1 = 0. # initializing with 0 
best_params = None 
for max_depth, min_child_weight in gridsearch_params:
    print("CV with max_depth={}, min_child_weight={}".format(
                             max_depth,
                             min_child_weight))

# Update our parameters
    params['max_depth'] = max_depth
    params['min_child_weight'] = min_child_weight

    
# Cross-validation
    cv_results = xgb.cv(        params,
        dtrain,        feval= custom_eval,
        num_boost_round=200,
        maximize=True,
        seed=16,
        nfold=5,
        early_stopping_rounds=10
    )    
    
# Finding best F1 Score
    
mean_f1 = cv_results['test-f1_score-mean'].max()

boost_rounds = cv_results['test-f1_score-mean'].idxmax() 

print("\tF1 Score {} for {} rounds".format(mean_f1, boost_rounds))

if mean_f1 > max_f1:
        max_f1 = mean_f1
        best_params = (max_depth,min_child_weight) 
        
print("Best params: {}, {}, F1 Score: {}".format(best_params[0], best_params[1], max_f1))

####  Updating max_depth and min_child_weight parameters

params['max_depth'] = 9 
params['min_child_weight'] = 7

### 4.2)-Tuning subsample and colsample

gridsearch_params = [
    (subsample, colsample)
    for subsample in [i/10. for i in range(5,10)]
    for colsample in [i/10. for i in range(5,10)] ]
max_f1 = 0. 
best_params = None 
for subsample, colsample in gridsearch_params:
    print("CV with subsample={}, colsample={}".format(
                             subsample,
                             colsample))
    # Update our parameters
    params['colsample'] = colsample
    params['subsample'] = subsample
    cv_results = xgb.cv(
        params,
        dtrain,
        feval= custom_eval,
        num_boost_round=200,
        maximize=True,
        seed=16,
        nfold=5,
        early_stopping_rounds=10
    )
     # Finding best F1 Score
    mean_f1 = cv_results['test-f1_score-mean'].max()
    boost_rounds = cv_results['test-f1_score-mean'].idxmax()
    print("\tF1 Score {} for {} rounds".format(mean_f1, boost_rounds))
    if mean_f1 > max_f1:
        max_f1 = mean_f1
        best_params = (subsample, colsample) 

print("Best params: {}, {}, F1 Score: {}".format(best_params[0], best_params[1], max_f1))

#### Updating subsample and colsample_bytree

params['subsample'] = .9 
params['colsample_bytree'] = .5

### 4.3)- tune the learning rate

max_f1 = 0. 
best_params = None 
for eta in [.3, .2, .1, .05, .01, .005]:
    print("CV with eta={}".format(eta))
     # Update ETA
    params['eta'] = eta

     # Run CV
    cv_results = xgb.cv(
        params,
        dtrain,
        feval= custom_eval,
        num_boost_round=1000,
        maximize=True,
        seed=16,
        nfold=5,
        early_stopping_rounds=20
    )

     # Finding best F1 Score
    mean_f1 = cv_results['test-f1_score-mean'].max()
    boost_rounds = cv_results['test-f1_score-mean'].idxmax()
    print("\tF1 Score {} for {} rounds".format(mean_f1, boost_rounds))
    if mean_f1 > max_f1:
        max_f1 = mean_f1
        best_params = eta 
print("Best params: {}, F1 Score: {}".format(best_params, max_f1))

#### Update learning rate

params['eta'] = .2

### 4.4)- List of tunned parameters

params
{'colsample': 0.9,
 'colsample_bytree': 0.5, 'eta': 0.2,
 'max_depth': 9, 'min_child_weight': 7,
 'objective': 'binary:logistic',
 'subsample': 0.9}

# 5)- Train the Tunned model

xgb_model = xgb.train(
    params,
    dtrain,
    feval= custom_eval,
    num_boost_round= 1000,
    maximize=True,
    evals=[(dvalid, "Validation")],
    early_stopping_rounds=10
 )

prediction = xgb_model.predict(dvalid)

This xgb_model can be tested on test set or out of sample data
